Add rpm-packaging-status.py for generating reports
The added tool can generate an overview about the current state of
packaging in the rpm-packaging project.
It uses the data from https://github.com/openstack/releases
and collects the needed versions for a given release and compares
this version with the version from the .spec.j2 templates.
The tool can generate a pretty table and also html output.

Change-Id: I8eb64c7bfeb82c9c87013000b71edf8f2354c652