Fix version parsing on python3
When run under python3, comparing a Version object with the string
obtained from parsing upper constraints caused a TypeError:

 TypeError: '>' not supported between instances of 'Version' and 'str'

This patch converts the upper constraints version number to a Version so
that it can be properly compared. It skips this if the string is '-'
because that means it is not in upper constraints.

Change-Id: Id1c97d42100cb973b856ecfbb8162ff30e94042d