Fix reporting when package version is unset
When a package tracks a stable branch rather than release, the spec file
doesn't contain a specific version number and so the status tool usually
reports it as "{{ py2rpmversion() }}" which gets converted to a Version
object that uses version 0.0. This causes the report to make it appear
like this packages need to be upgraded, but in reality no changes are
needed because they automatically track the branch. This change detects
versions like this and reports them as "ok" instead of "needs upgrade".

Change-Id: Ia1191c6c0938e78d8824cb8d893d561ddd99de4a